Add GitHub build workflows. Refactor weather forecast service into ArcadeClientService.
