+++ src/Blazor.Arcade.Client/Pages/FetchData.razor
@@ -1,6 +1,6 @@
 ﻿@page "/fetchdata"
 @attribute [Authorize]
-@inject WeatherForecastHttpClient Http
+@inject ArcadeClientService Http
 
 <PageTitle>Weather forecast</PageTitle>
 
@@ -42,7 +42,6 @@ else
 
     protected override async Task OnInitializedAsync()
     {
-        // forecasts = await Http.GetFromJsonAsync<WeatherForecast[]>("sample-data/weather.json");
         forecasts = await Http.GetWeatherForecastsAsync();
     }
