+++ src/Blazor.Arcade.Client/_Imports.razor
@@ -10,5 +10,5 @@
 @using Microsoft.JSInterop
 
 @using Blazor.Arcade.Client
-@using Blazor.Arcade.Client.Client
+@using Blazor.Arcade.Client.Services
 @using Blazor.Arcade.Client.Shared
